migration to change Pet xy pos from Int to Float. update alarm to correct time in mainView after set

+++ Helpers/StateMachine.swift
@@ -141,8 +141,8 @@ class StateMachine {
         var pet = getRealmPet()
         if let pet = pet {
             realm.write{
-                pet.x = Int(x)
-                pet.y = Int(y)
+                pet.x = Float(x)
+                pet.y = Float(y)
             }
         }
     }
